+++ plugin-server/src/index.ts
@@ -1,7 +1,7 @@
 import { getPluginServerCapabilities } from './capabilities'
 import { defaultConfig } from './config/config'
 import { initApp } from './init'
-import { startPluginsServer } from './main/pluginsServer'
+import { startPluginsServer } from './server'
 
 initApp(defaultConfig)
 const capabilities = getPluginServerCapabilities(defaultConfig)
