+++ plugin-server/src/server.ts
@@ -8,50 +8,50 @@ import * as schedule from 'node-schedule'
 import { Counter } from 'prom-client'
 import v8Profiler from 'v8-profiler-next'
 
-import { getPluginServerCapabilities } from '../capabilities'
-import { CdpApi } from '../cdp/cdp-api'
-import { CdpCyclotronWorkerPlugins } from '../cdp/consumers/cdp-cyclotron-plugins-worker.consumer'
-import { CdpCyclotronWorker, CdpCyclotronWorkerFetch } from '../cdp/consumers/cdp-cyclotron-worker.consumer'
-import { CdpInternalEventsConsumer } from '../cdp/consumers/cdp-internal-event.consumer'
-import { CdpProcessedEventsConsumer } from '../cdp/consumers/cdp-processed-events.consumer'
-import { defaultConfig } from '../config/config'
+import { getPluginServerCapabilities } from './capabilities'
+import { CdpApi } from './cdp/cdp-api'
+import { CdpCyclotronWorkerPlugins } from './cdp/consumers/cdp-cyclotron-plugins-worker.consumer'
+import { CdpCyclotronWorker, CdpCyclotronWorkerFetch } from './cdp/consumers/cdp-cyclotron-worker.consumer'
+import { CdpInternalEventsConsumer } from './cdp/consumers/cdp-internal-event.consumer'
+import { CdpProcessedEventsConsumer } from './cdp/consumers/cdp-processed-events.consumer'
+import { defaultConfig } from './config/config'
 import {
     KAFKA_EVENTS_PLUGIN_INGESTION,
     KAFKA_EVENTS_PLUGIN_INGESTION_HISTORICAL,
     KAFKA_EVENTS_PLUGIN_INGESTION_OVERFLOW,
-} from '../config/kafka-topics'
-import { IngestionConsumer } from '../ingestion/ingestion-consumer'
-import { KafkaProducerWrapper } from '../kafka/producer'
-import { Hub, PluginServerCapabilities, PluginServerService, PluginsServerConfig } from '../types'
-import { closeHub, createHub, createKafkaClient } from '../utils/db/hub'
-import { PostgresRouter } from '../utils/db/postgres'
-import { createRedisClient } from '../utils/db/redis'
-import { cancelAllScheduledJobs } from '../utils/node-schedule'
-import { captureException } from '../utils/posthog'
-import { flush as posthogFlush, shutdown as posthogShutdown } from '../utils/posthog'
-import { PubSub } from '../utils/pubsub'
-import { status } from '../utils/status'
-import { delay } from '../utils/utils'
-import { ActionManager } from '../worker/ingestion/action-manager'
-import { ActionMatcher } from '../worker/ingestion/action-matcher'
-import { AppMetrics } from '../worker/ingestion/app-metrics'
-import { GroupTypeManager } from '../worker/ingestion/group-type-manager'
-import { OrganizationManager } from '../worker/ingestion/organization-manager'
-import { TeamManager } from '../worker/ingestion/team-manager'
-import { teardownPlugins } from '../worker/plugins/teardown'
-import { RustyHook } from '../worker/rusty-hook'
-import { initPlugins as _initPlugins, reloadPlugins } from '../worker/tasks'
-import { syncInlinePlugins } from '../worker/vm/inline/inline'
-import { populatePluginCapabilities } from '../worker/vm/lazy'
+} from './config/kafka-topics'
+import { IngestionConsumer } from './ingestion/ingestion-consumer'
+import { KafkaProducerWrapper } from './kafka/producer'
 import {
     startAsyncOnEventHandlerConsumer,
     startAsyncWebhooksHandlerConsumer,
-} from './ingestion-queues/on-event-handler-consumer'
-import { SessionRecordingIngester } from './ingestion-queues/session-recording/session-recordings-consumer'
-import { DefaultBatchConsumerFactory } from './ingestion-queues/session-recording-v2/batch-consumer-factory'
-import { SessionRecordingIngester as SessionRecordingIngesterV2 } from './ingestion-queues/session-recording-v2/consumer'
-import { expressApp, setupCommonRoutes } from './services/http-server'
-import { getObjectStorage } from './services/object_storage'
+} from './main/ingestion-queues/on-event-handler-consumer'
+import { SessionRecordingIngester } from './main/ingestion-queues/session-recording/session-recordings-consumer'
+import { DefaultBatchConsumerFactory } from './main/ingestion-queues/session-recording-v2/batch-consumer-factory'
+import { SessionRecordingIngester as SessionRecordingIngesterV2 } from './main/ingestion-queues/session-recording-v2/consumer'
+import { expressApp, setupCommonRoutes } from './main/services/http-server'
+import { getObjectStorage } from './main/services/object_storage'
+import { Hub, PluginServerCapabilities, PluginServerService, PluginsServerConfig } from './types'
+import { closeHub, createHub, createKafkaClient } from './utils/db/hub'
+import { PostgresRouter } from './utils/db/postgres'
+import { createRedisClient } from './utils/db/redis'
+import { cancelAllScheduledJobs } from './utils/node-schedule'
+import { captureException } from './utils/posthog'
+import { flush as posthogFlush, shutdown as posthogShutdown } from './utils/posthog'
+import { PubSub } from './utils/pubsub'
+import { status } from './utils/status'
+import { delay } from './utils/utils'
+import { ActionManager } from './worker/ingestion/action-manager'
+import { ActionMatcher } from './worker/ingestion/action-matcher'
+import { AppMetrics } from './worker/ingestion/app-metrics'
+import { GroupTypeManager } from './worker/ingestion/group-type-manager'
+import { OrganizationManager } from './worker/ingestion/organization-manager'
+import { TeamManager } from './worker/ingestion/team-manager'
+import { teardownPlugins } from './worker/plugins/teardown'
+import { RustyHook } from './worker/rusty-hook'
+import { initPlugins as _initPlugins, reloadPlugins } from './worker/tasks'
+import { syncInlinePlugins } from './worker/vm/inline/inline'
+import { populatePluginCapabilities } from './worker/vm/lazy'
 
 CompressionCodecs[CompressionTypes.Snappy] = SnappyCodec
 CompressionCodecs[CompressionTypes.LZ4] = new LZ4().codec
